testes-automatizados-eps-mds/ta-eps-mds-back-end/ta-eps-mds-django/djangoProject/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
psycopg2==2.9.3
djangorestframework==3.12.4
django-cors-headers